accurately set trl in bnb qpeft fix and file rename
